move AnalyticsFollowable to its own file, and follow 2 meters behind the chest joint instead of directly at the neck joint
