+++ content_scripts/CameraVis.js
@@ -23,9 +23,12 @@ function setMatrixFromArray(matrix, array) {
 }
 
 export class CameraVis extends Followable {
+    static count = 0;
+
     constructor(id, floorOffset, color) {
+        CameraVis.count++;
         let parentNode = realityEditor.sceneGraph.getVisualElement('CameraGroupContainer');
-        let displayName = `Live Video ${id}`;
+        let displayName = `Live Video ${CameraVis.count}`;
         super('CameraVisFollowable_' + id, displayName, parentNode);
 
         this.id = id;
@@ -49,14 +52,6 @@ export class CameraVis extends Followable {
         this.container.add(this.phone);
 
         this.maxDepthMeters = 5; // this goes down if lidar is pointed at a wall/floor/object closer than 5 meters
-
-        // this.followable = new Followable('CameraVis_' + id);
-        // this.followable.enableFirstPersonMode()
-
-        // let parentNode = realityEditor.sceneGraph.getGroundPlaneNode();
-        // let parentNode = realityEditor.sceneGraph.getSceneNodeById(elementId);
-        let sceneGraphNodeId = realityEditor.sceneGraph.addVisualElement('CameraVis_' + id, parentNode);
-        this.sceneGraphNode = realityEditor.sceneGraph.getSceneNodeById(sceneGraphNodeId);
 
         this.cameraMeshGroup = new THREE.Group();
 
@@ -337,14 +332,6 @@ export class CameraVis extends Followable {
         if (addToHistory) {
             this.historyPoints.push(nextHistoryPoint);
             this.historyMesh.setPoints(this.historyPoints);
-        }
-
-        if (this.sceneGraphNode) {
-            this.sceneGraphNode.setLocalMatrix(newMatrix);
-        }
-
-        if (this.sceneNode) {
-            this.sceneNode.setLocalMatrix(newMatrix);
         }
 
         if (this.firstPersonMode) {
